Location test

+++ Quickblox_8.1/Quickblox.Sdk/Modules/LocationModule/Models/Location.cs
@@ -5,10 +5,10 @@ namespace Quickblox.Sdk.Modules.LocationModule.Models
     public class Location
     {
         [JsonProperty(PropertyName = "latitude")]
-        public float Latitude { get; set; }
+        public double Latitude { get; set; }
 
         [JsonProperty(PropertyName = "longitude")]
-        public float Longtitude { get; set; }
+        public double Longtitude { get; set; }
 
         [JsonProperty(PropertyName = "status")]
         public string Status { get; set; }
